+++ src/bigxml/exceptions.py
@@ -6,7 +6,7 @@ from defusedxml.ElementTree import ParseError
 from bigxml.typing import T
 
 
-class BigXmlException(ValueError):
+class BigXmlError(ValueError):
     """Error parsing XML content"""
 
     def __init__(self, msg: str, security: bool) -> None:
@@ -19,7 +19,7 @@ def rewrite_exceptions(iterable: Iterable[T]) -> Iterator[T]:
     try:
         yield from iterable
     except ParseError as ex:
-        raise BigXmlException(str(ex), False) from ex
+        raise BigXmlError(str(ex), False) from ex
     except DefusedXmlException as ex:
         # defusedxml has usable wording in it's exception's doc
         # except for base DefusedXmlException
@@ -27,4 +27,4 @@ def rewrite_exceptions(iterable: Iterable[T]) -> Iterator[T]:
         # pylint: disable-next=unidiomatic-typecheck
         if not msg or type(ex) is DefusedXmlException:
             msg = "Invalid XML"
-        raise BigXmlException(msg, True) from ex
+        raise BigXmlError(msg, True) from ex